CustomDonationAmount component
