+++ forms/CustomDonationAmount/index.js
@@ -1,33 +1,32 @@
+'use strict'
+
 import React from 'react'
-import { shallow } from 'enzyme'
-import { expect } from 'chai'
 
-import CustomDonationAmount from '../'
+const CustomDonationAmount = ({
+  handleChanged
+}) => {
+  return (
+    <li
+      className='string required'
+      id='donation_donation_builder_defined_amount_input'
+    >
+      <label for='donation_donation_builder_defined_amount'>Other amount</label>
+      <input
+        className='surcharge-input'
+        id='donation_donation_builder_defined_amount'
+        name='defined_amount'
+        size='30'
+        type='number'
+        onChange={function (event) {
+          handleChanged(event.target.value)
+        }}
+      />
+    </li>
+  )
+}
 
-describe('CustomDonationAmount display component', () => {
-  let wrapper, changeStub
+CustomDonationAmount.propTypes = {
+  handleChanged: React.PropTypes.func.isRequired
+}
 
-  beforeEach(() => {
-    changeStub = sinon.stub()
-    wrapper = shallow(<CustomDonationAmount handleChanged={changeStub } />)
-  })
-
-  it('should render a defined_amount number input', () => {
-    const numInput = wrapper.find('input[type="number"]')
-    expect(numInput).to.be.present()
-    expect(numInput).to.have.attr('name', 'defined_amount')
-  })
-
-  it('should render a matching label', () => {
-    const label = wrapper.find('label')
-    expect(label).to.be.present()
-    expect(label).to.have.text('Other amount')
-  })
-
-  it('should call handleChanged prop when input value changes', () => {
-    const numInput = wrapper.find('input[type="number"]')
-    numInput.simulate('input', { target: { value: '35' }})
-    expect(changeStub).to.be.calledWith('35')
-  })
-})
-
+export default CustomDonationAmount
